+++ app/Http/Controllers/Api/v2/ProductController.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers\Api\v2;
 use App\Enums\RouteGroupEnum;
 use App\Http\Requests\v2\Product\IndexRequest;
 use App\Http\Resources\Product\IndexResource;
-use App\Services\Elasticsearch\Contracts\ElasticsearchServiceContract;
 use App\Services\Elasticsearch\PaginationRequestMapper;
+use App\Services\Elasticsearch\Repositories\Contracts\ElasticsearchRepositoryContract;
 use Dedoc\Scramble\Attributes\Endpoint;
 use Dedoc\Scramble\Attributes\Group;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
@@ -15,7 +15,7 @@ use Illuminate\Support\Arr;
 
 class ProductController extends Controller
 {
-    public function __construct(private readonly ElasticsearchServiceContract $searchService) {}
+    public function __construct(private readonly ElasticsearchRepositoryContract $repository) {}
 
     #[Group(
         name: RouteGroupEnum::PRODUCT->value,
@@ -26,7 +26,7 @@ class ProductController extends Controller
     {
         $inputData = $request->validated();
 
-        $searchResult = $this->searchService->findInSearchIndex($mapper->map(
+        $searchResult = $this->repository->findInSearchIndex($mapper->map(
             Arr::get($inputData, 'search'),
             Arr::get($inputData, 'per_page'),
             Arr::get($inputData, 'sorted_by'),
