+++ app/Services/Elasticsearch/Repositories/ProductElasticsearchRepository.php
@@ -2,25 +2,27 @@
 
 declare(strict_types=1);
 
-namespace App\Services\Elasticsearch;
+namespace App\Services\Elasticsearch\Repositories;
 
 use App\Clients\Elasticsearch\Contracts\ElasticsearchClientContract;
+use App\Enums\SortedByEnum;
 use App\Events\Elasticsearch\SearchIndexFilledEvent;
-use App\Services\Elasticsearch\Abstract\BaseElasticsearchService;
 use App\Services\Elasticsearch\Entities\BulkIndexResult;
 use App\Services\Elasticsearch\Factories\BulkIndexResultFactory;
 use App\Services\Elasticsearch\Factories\SearchResultFactory;
-use App\Services\User\Entities\UserEnriched;
-use App\Services\User\UserService;
+use App\Services\Elasticsearch\Repositories\Abstract\BaseElasticsearchRepository;
+use App\Services\Product\Dto\FilterDto;
+use App\Services\Product\Entities\ProductEnriched;
+use App\Services\Product\ProductService;
 use Illuminate\Contracts\Events\Dispatcher;
 
-class UsersIndexElasticsearchService extends BaseElasticsearchService
+class ProductElasticsearchRepository extends BaseElasticsearchRepository
 {
     public function __construct(
         protected ElasticsearchClientContract $client,
         protected SearchResultFactory $searchResultFactory,
         private readonly Dispatcher $dispatcher,
-        private readonly UserService $userService,
+        private readonly ProductService $productService,
         private readonly BulkIndexResultFactory $bulkIndexResultFactory,
     ) {
         parent::__construct($client, $searchResultFactory);
@@ -28,7 +30,7 @@ class UsersIndexElasticsearchService extends BaseElasticsearchService
 
     protected function indexName(): string
     {
-        return 'users';
+        return 'products';
     }
 
     /**
@@ -64,19 +66,7 @@ class UsersIndexElasticsearchService extends BaseElasticsearchService
                         'type' => 'text',
                         'analyzer' => 'my_ngram_analyzer',
                     ],
-                    'email' => [
-                        'type' => 'text',
-                        'analyzer' => 'my_ngram_analyzer',
-                    ],
-                    'reserve_email' => [
-                        'type' => 'text',
-                        'analyzer' => 'my_ngram_analyzer',
-                    ],
-                    'phone' => [
-                        'type' => 'text',
-                        'analyzer' => 'my_ngram_analyzer',
-                    ],
-                    'telegram' => [
+                    'description' => [
                         'type' => 'text',
                         'analyzer' => 'my_ngram_analyzer',
                     ],
@@ -91,23 +81,23 @@ class UsersIndexElasticsearchService extends BaseElasticsearchService
     protected function multiMatchFieldsSettings(): array
     {
         return [
-            'name^5',
-            'email^4',
-            'reserve_email^3',
-            'telegram^2',
-            'phone^1',
+            'name^2',
+            'description^1',
         ];
     }
 
     public function fillSearchIndex(?int $limit = null): ?BulkIndexResult
     {
-        $users = $this->userService->getList($limit);
-        if ($users->isEmpty()) {
+        $products = $this->productService->getList(new FilterDto(
+            sortedBy: SortedByEnum::ASC,
+            limit: $limit
+        ));
+        if ($products->isEmpty()) {
             return null;
         }
 
-        $body = $users->map(fn (UserEnriched $user): string => $this->makeDocElement(
-            $user->toArray(),
+        $body = $products->map(fn (ProductEnriched $product): string => $this->makeDocElement(
+            $product->toArray(),
             $this->indexName()
         ))
             ->implode('');
@@ -116,7 +106,7 @@ class UsersIndexElasticsearchService extends BaseElasticsearchService
 
         return tap(
             $this->bulkIndexResultFactory->make($result),
-            fn (): ?array => $this->dispatcher->dispatch(new SearchIndexFilledEvent($users, $this->indexName()))
+            fn (): ?array => $this->dispatcher->dispatch(new SearchIndexFilledEvent($products, $this->indexName()))
         );
     }
 }
